PaymentRequest moved to internal package, completeCheckout to return Payment object
